+++ src/modules/auth/auth.controller.ts
@@ -1,8 +1,9 @@
 import { Body, Controller, Inject, Post } from '@nestjs/common';
-import { LoginAccountRequest, RegisterAccountRequest } from '@/modules/auth/dtos';
+
 import { AuthService } from '@/modules/auth/auth.service';
 import { UserService } from '@/modules/user/user.service';
 import { RequestOk } from '@/utils';
+import { LoginAccountRequest, RegisterAccountRequest } from '@/request';
 
 @Controller('/api')
 export class AuthController {
