+++ src/modules/bot/service.ts
@@ -2,8 +2,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { pick } from 'lodash';
-import { CreateBotRequest, ListBotsRequest, UpdateBotRequest } from './dtos';
 import { generateId } from './utils';
+import { CreateBotRequest, ListBotsRequest, UpdateBotRequest } from '@/request';
 import { Bot } from '@/entity';
 import { RequestFail } from '@/utils';
 import { BotRunService } from '@/modules/botrun/services';
