--- src/modules/user/dtos/index.ts
@@ -1,12 +1,0 @@
-import { IsOptional, IsString, Length, Matches } from 'class-validator';
-
-export class UpdateUserRequest {
-  @IsOptional()
-  @IsString()
-  @Length(6, 18)
-  @Matches(/^\w*$/)
-    Name?: string;
-  @IsOptional()
-  @IsString()
-    Avatar?: string;
-}
