+++ src/request/index.ts
@@ -1,0 +1,6 @@
+export * from './auth';
+export * from './user';
+export * from './game';
+export * from './bot';
+export * from './rate';
+export * from './tape';
