+++ src/request/rate.ts
@@ -4,12 +4,12 @@ import { Rate } from '@/entity/rate';
 
 
 export class OnlyRateId {
-  @IsNotEmpty()
-    UserId: string;
-  @IsNotEmpty()
-    GameId: string;
-  @IsOptional()
-    BotId = '';
+    @IsNotEmpty()
+      UserId: string;
+    @IsNotEmpty()
+      GameId: string;
+    @IsOptional()
+      BotId = '';
 }
 
 export class CreateRateRequest extends OnlyRateId {
@@ -19,25 +19,25 @@ export class CreateRateRequest extends OnlyRateId {
 export class GetRateRequest extends OnlyRateId {}
 
 export interface ListRatesFilter {
-  UserIds?: string[];
-  GameIds?: string[];
-  BotIds?: string[];
+    UserIds?: string[];
+    GameIds?: string[];
+    BotIds?: string[];
 }
 
 export class ListRatesRequest extends CommonListRequest<ListRatesFilter> {
-  @IsOptional()
-  @IsBoolean()
-    WithRank?: boolean;
+    @IsOptional()
+    @IsBoolean()
+      WithRank?: boolean;
 }
 
 export interface WithRankRate extends Rate {
-  Rank?: number;
+    Rank?: number;
 }
 
 export interface ListRatesResponse extends CommonListResponse<WithRankRate> {}
 
 export class UpdateRateRequest extends OnlyRateId {
-  @IsOptional()
-  @IsInt()
-    Score?: number;
+    @IsOptional()
+    @IsInt()
+      Score?: number;
 }
